fix(clippy): replace deprecated cargo_bin with cargo_bin_cmd macro
driver/tests/dynamic_driver_lsm_packaging.rs:
- Command::cargo_bin() deprecated in assert_cmd 2.1.0 (incompatible with
  custom cargo build-dir; replacement: cargo_bin_cmd! macro)
- 2 sites updated, import replaced
- simple-driver clippy warnings: 2 → 0 (real fix, no allow added)

+++ src/compiler_rust/runtime/src/value/ffi/concurrent/arena.rs
@@ -278,7 +278,7 @@ mod tests {
 
             let ptr = rt_arena_alloc(handle, 32, align);
             assert_ne!(ptr, 0);
-            assert_eq!(ptr % (align as i64), 0);
+            assert_eq!(ptr % align, 0);
         }
 
         rt_arena_free(handle);
